rework package structure add fill_arguments to RndCodeInstr - inserts pushcodes for instr arguments setup.py use find_packages
