+++ evmlab/tools/statetests/rndval/code.py
@@ -1,0 +1,351 @@
+
+
+import random
+import binascii
+from .base import _RndBase, WeightedRandomizer
+
+
+class RndCodeBytes(_RndBase):
+    """
+    Random bytecode (could be empty string)
+    """
+    placeholder = "[CODEBYTES]"
+
+    # analyzed based on statedump.json
+    LIKELYHOOD_BY_OPCODE_INT = {0: 0.07243246371171248, 1: 0.053453584934496264, 2: 0.021083348438644005,
+                                3: 0.017330067476311318, 4: 0.013420068915758572, 5: 0.003349158367631274,
+                                6: 0.002551575750467376, 7: 0.0016572525039273808, 8: 0.0025380809324774844,
+                                9: 0.00241212929790516, 10: 0.01017203838170587, 11: 0.0014590508355902445,
+                                12: 0.0012005389930266777, 13: 0.0009475250392738078, 14: 0.0009837888917322414,
+                                15: 0.0017069001059148842, 16: 0.0035604105801150137, 17: 0.0026359877806922142,
+                                18: 0.0010751982349888754, 19: 0.0008925461511668408, 20: 0.008195686189836435,
+                                21: 0.012244686929108111, 22: 0.014358486341244961, 23: 0.0021460092657752754,
+                                24: 0.0006377551020408163, 25: 0.0030695435175114977, 26: 0.000601824454964849,
+                                27: 0.00036919156377284454, 28: 0.00043288932605435067, 29: 0.0008244056504524485,
+                                30: 0.00036130570305447154, 31: 0.0030557154941391394, 32: 0.023564062511106845,
+                                33: 0.0004505492113250734, 34: 0.0005750014216762985, 35: 0.0005608957271518848,
+                                36: 0.0015385758535389077, 37: 0.00042333743842364535, 38: 0.0002535692960569808,
+                                39: 0.0008227396235401162, 40: 0.0006957328385899814, 41: 0.0004188946999907591,
+                                42: 0.0003723014806758649, 43: 0.0006982318789584799, 44: 0.00029227665465350193,
+                                45: 0.00048559131071446344, 46: 0.0004050111423879897, 47: 0.00044638414404424254,
+                                48: 0.0007944727002608776, 49: 0.0008257384719823144, 50: 0.001523803748249561,
+                                51: 0.0028978872113108564, 52: 0.001212867592177937, 53: 0.004420080466878497,
+                                54: 0.0017790946054492853, 55: 0.0013380417475245062, 56: 0.0004898674464561164,
+                                57: 0.0009052079557005665, 58: 0.0004485499790302746, 59: 0.0003160453052694432,
+                                60: 0.00045471427860590424, 61: 0.0008655009809566459, 62: 0.0002909438331236361,
+                                63: 0.0002732284136225023, 64: 0.011408674624499748, 65: 0.0008378449342119293,
+                                66: 0.000654637508085784, 67: 0.0007486569601717384, 68: 0.0011990951030359897,
+                                69: 0.0008550605456393634, 70: 0.0005191339858827543, 71: 0.0005135805628416466,
+                                72: 0.0008782738539511939, 73: 0.0009705162106639939, 74: 0.00022680179699884133,
+                                75: 0.0004071769773740217, 76: 0.0004890344329999502, 77: 0.00024068535460161075,
+                                78: 0.00041228612657184083, 79: 0.0007263877337768963, 80: 0.046506696983913735,
+                                81: 0.014441176810327057, 82: 0.021314981713688612, 83: 0.0004973645675616119,
+                                84: 0.013179161424235315, 85: 0.005296632759686947, 86: 0.013287897447380206,
+                                87: 0.016262088691275885, 88: 0.0003118802379886124, 89: 0.0008521727656579873,
+                                90: 0.002300116755166016, 91: 0.023078249063470738, 92: 0.0004498272663297294,
+                                93: 0.00035830685461227333, 94: 0.0007062288081376751, 95: 0.000522410505477008,
+                                96: 0.10626469435736677, 97: 0.039956712178079175, 98: 0.000965629198387819,
+                                99: 0.007088167032748314, 100: 0.0009970615728004892, 101: 0.0010221630449462962,
+                                102: 0.0005090822901783493, 103: 0.0008575595860078618, 104: 0.0013230475053135151,
+                                105: 0.0008250720612173814, 106: 0.0002661755663602954, 107: 0.0006463073735241223,
+                                108: 0.0006832931709779, 109: 0.0006960105097420368, 110: 0.0007110047519530278,
+                                111: 0.0005494001414567917, 112: 0.00040462240277511213, 113: 0.000416673330774316,
+                                114: 0.000564616520589427, 115: 0.005663991693856226, 116: 0.0010490971466956688,
+                                117: 0.00047431786194101464, 118: 0.0005563419202581764, 119: 0.0005421806915033516,
+                                120: 0.0004955874721884574, 121: 0.00041184185272855224, 122: 0.0007298863902927942,
+                                123: 0.00025245861144875925, 124: 0.0007919736598923791, 125: 0.00046454383738866496,
+                                126: 0.000488867830308717, 127: 0.0020719266024068978, 128: 0.02899347761570668,
+                                129: 0.03278107873243341, 130: 0.01796238022377185, 131: 0.010644412545582497,
+                                132: 0.006120038793991996, 133: 0.0035153723192516296, 134: 0.0019657451538609176,
+                                135: 0.0029229886834566637, 136: 0.001234248270886202, 137: 0.0006209282302262598,
+                                138: 0.0007849763468605833, 139: 0.0005862193362193362, 140: 0.0004004573354942813,
+                                141: 0.0007846986757085279, 142: 0.0003771329587216287, 143: 0.00029555317424775554,
+                                144: 0.043290098824273704, 145: 0.017430473364894547, 146: 0.0070315776519594255,
+                                147: 0.0030437756346007577, 148: 0.0019938454744489226, 149: 0.0014170114231690587,
+                                150: 0.000834124140774387, 151: 0.000562506219833806, 152: 0.0005322955984901797,
+                                153: 0.0003641379488054365, 154: 0.0004439406379061551, 155: 0.00044988280056014045,
+                                156: 0.0003082149787814812, 157: 0.00028750071083814924, 158: 0.0002635099233005637,
+                                159: 0.0004061773612266223, 160: 0.004447958650544858, 161: 0.0008614469821366373,
+                                162: 0.000475261943858003, 163: 0.00048481383148870835, 164: 0.00040706590891319955,
+                                165: 0.00035975074460296136, 166: 0.0003012731999800965, 167: 0.00043155650452448484,
+                                168: 0.0003633049353492703, 169: 0.0005163572743622004, 170: 0.0007329407729654036,
+                                171: 0.0003227649471491836, 172: 0.00030571593841298276, 173: 0.0003603060869070721,
+                                174: 0.0007914738518186794, 175: 0.00028605682084746126, 176: 0.0005365717342318328,
+                                177: 0.0002877228477597936, 178: 0.0004096760177425202, 179: 0.0002733950163137355,
+                                180: 0.00045865720896509074, 181: 0.000606045056476091, 182: 0.0006021021261169044,
+                                183: 0.0005261868331449612, 184: 0.00030965886877216926, 185: 0.0003835749294493137,
+                                186: 0.00021675010129443628, 187: 0.000447383760191642, 188: 0.00031121382722367944,
+                                189: 0.0002313556038925497, 190: 0.0007057845342943865, 191: 0.0002746723036131903,
+                                192: 0.0009628524868672652, 193: 0.0005016962375336759, 194: 0.0005437911841852729,
+                                195: 0.00033381625900098805, 196: 0.00045088241670753984, 197: 0.00032904031518563536,
+                                198: 0.0006327014870734082, 199: 0.00021436212938675994, 200: 0.00048403635226295326,
+                                201: 0.0007530441643742136, 202: 0.0004883124880046062, 203: 0.00030965886877216926,
+                                204: 0.0009209241429069015, 205: 0.0004239483149581672, 206: 0.00025695688411205654,
+                                207: 0.00047809418960896794, 208: 0.0002896665458241813, 209: 0.0005776115305056192,
+                                210: 0.0006451966889159008, 211: 0.00032565272713055965, 212: 0.00026040000639754334,
+                                213: 0.00036341600381009246, 214: 0.0006499170985008423, 215: 0.0004454955963576653,
+                                216: 0.000258012034489867, 217: 0.0003790211225556053, 218: 0.0013524251132009753,
+                                219: 0.00045665797667029197, 220: 0.0004726518350286823, 221: 0.00029688599577762137,
+                                222: 0.0002613440883145317, 223: 0.00025884504794603316, 224: 0.0015590679845605953,
+                                225: 0.000990564067842393, 226: 0.0003195994960157522, 227: 0.00046082304395112276,
+                                228: 0.000388795147107955, 229: 0.0005061389759665622, 230: 0.0003898502974857655,
+                                231: 0.0003013287342105076, 232: 0.0007938618237263558, 233: 0.000223691880095821,
+                                234: 0.00040889853851676513, 235: 0.0005472343064707596, 236: 0.0002690633463416715,
+                                237: 0.00025368036451780293, 238: 0.0009040972710923449, 239: 0.00032454204252233806,
+                                240: 0.0004570467162831695, 241: 0.0019200404822326005, 242: 0.0014862070742612614,
+                                243: 0.0020987496356954486, 244: 0.00025712348680328973, 245: 0.0006651334776334776,
+                                246: 0.0008061348886472039, 247: 0.0003128243199056007, 248: 0.0006469737842890553,
+                                249: 0.00037702189026080654, 250: 0.000450271540173018, 251: 0.0003425351331755273,
+                                252: 0.0007233333511042871, 253: 0.0003427017358667605, 254: 0.00037085759068517687,
+                                255: 0.08685514762331265}
+    LIKELYHOOD_PROLOG_BY_OPCODE_INT = {0: 0.2926344336089935, 1: 0.023987953991816913, 2: 0.009097574242985644,
+                                       3: 0.005486925193343161, 4: 0.0176543394798458, 5: 0.0004651272793207565,
+                                       6: 0.0018478954961489039, 7: 0.00036106490496424826, 8: 0.0005455391140507856,
+                                       9: 0.0015719725338399805, 10: 0.009316735910191018, 11: 0.00022862188305596505,
+                                       12: 0.0002743462596671581, 13: 0.0005723430589607952, 14: 0.00043516992912721625,
+                                       15: 0.0009602119088350532, 16: 0.00039890576836661487, 17: 0.0007773144023902812,
+                                       18: 0.0016271571263017652, 19: 0.0013874983247534431, 20: 0.030185972076596214,
+                                       21: 0.013198577814217127, 22: 0.001466333456841707, 23: 0.0012629388160539862,
+                                       24: 0.000496661332156062, 25: 0.004160918271618564, 26: 0.0006937491623767216,
+                                       27: 0.00035633479703895245, 28: 0.00017659069587771094,
+                                       29: 0.00036737171553130935, 30: 0.003981174170457322, 31: 0.00035948820232248297,
+                                       32: 0.012192641528770882, 33: 0.00018762761437006787, 34: 0.004154611461051503,
+                                       35: 0.004458915070912201, 36: 0.00265359054609096, 37: 0.004241330106348593,
+                                       38: 0.0005786498695278564, 39: 0.00496661332156062, 40: 0.00016240037210182346,
+                                       41: 0.0006921724597349563, 42: 0.0002727695570253928, 43: 0.000342144473263065,
+                                       44: 0.00020181793814595537, 45: 0.0001324430219082832, 46: 0.000496661332156062,
+                                       47: 0.00024754231475714836, 48: 0.0040789297342467695, 49: 0.001917270412386576,
+                                       50: 0.0048798946762635305, 51: 0.0004998147374395926, 52: 0.004014284925934393,
+                                       53: 0.0182582165916419, 54: 0.01658533508872894, 55: 0.004274440861825664,
+                                       56: 0.001381191514186382, 57: 0.004089966652739127, 58: 0.00016082366946005818,
+                                       59: 0.00037525522874013573, 60: 0.0017769438772694663,
+                                       61: 0.00022389177513066922, 62: 0.0002711928543836275,
+                                       63: 0.00034845128383012607, 64: 0.016850221132545508, 65: 0.006896497355081318,
+                                       66: 0.00035633479703895245, 67: 0.00019078101965359844,
+                                       68: 0.0030840303672928803, 69: 0.0003437211759048302, 70: 0.00018132080380300677,
+                                       71: 0.0014931374017517166, 72: 0.0017422564191506302, 73: 0.00029011328608481084,
+                                       74: 9.30254558641513e-05, 75: 0.00022862188305596505, 76: 0.0005723430589607952,
+                                       77: 0.00020970145135478175, 78: 0.0014347994040064013,
+                                       79: 0.00046197387403722597, 80: 0.0003910222551577885,
+                                       81: 0.00024911901739891365, 82: 0.016810803566501376, 83: 0.0006212208408555188,
+                                       84: 0.000810425157867352, 85: 0.00034845128383012607, 86: 0.0008340756974938311,
+                                       87: 0.029941583167122598, 88: 0.0004950846295142967, 89: 0.0040600093025455865,
+                                       90: 0.008138939036792357, 91: 0.004080506436888535, 92: 0.00017501399323594568,
+                                       93: 0.0004241330106348593, 94: 0.00036106490496424826, 95: 0.0002932666913683414,
+                                       96: 0.11569213304216891, 97: 0.037607511411385366, 98: 0.0004257097132766246,
+                                       99: 0.0317484843945856, 100: 7.725842944649854e-05, 101: 0.0008183086710761784,
+                                       102: 0.0004067892815754413, 103: 0.0009728255299691755,
+                                       104: 0.0004603971713954607, 105: 0.0022862188305596506,
+                                       106: 0.0002601559358912706, 107: 0.0003437211759048302, 108: 0.0015120578334529,
+                                       109: 0.0006133373276466925, 110: 0.0006937491623767216,
+                                       111: 0.0010958083360268671, 112: 0.0008183086710761784,
+                                       113: 0.00014032653511710958, 114: 0.00019078101965359844,
+                                       115: 0.0045614007426269445, 116: 0.0005581527351849078,
+                                       117: 0.00010248567171474296, 118: 0.0002979967992936372,
+                                       119: 0.00068428894652613, 120: 0.0001513634536094665, 121: 0.000525041979707837,
+                                       122: 0.0014284925934393404, 123: 0.0008908369925973811,
+                                       124: 0.009488596498143433, 125: 0.00010879248228180406,
+                                       126: 0.0008167319684344131, 127: 0.0004872011163054704,
+                                       128: 0.027579682609758214, 129: 0.008786963822557885, 130: 0.0013291603270081278,
+                                       131: 0.00013717312983357904, 132: 0.0006543315963325896,
+                                       133: 0.00022073836984713867, 134: 0.00017816739851947623,
+                                       135: 0.004045818978769699, 136: 0.0010153965012968378,
+                                       137: 0.00023019858569773034, 138: 0.0008246154816432395,
+                                       139: 0.0009255244507162172, 140: 7.410502416296798e-05,
+                                       141: 0.0005770731668860911, 142: 0.0003121871230695247,
+                                       143: 0.0005061215480066537, 144: 0.00972667859704999,
+                                       145: 0.00019393442493712899, 146: 0.00030903371778599417,
+                                       147: 0.00019551112757889425, 148: 0.00028695988080128026,
+                                       149: 0.0010311635277144907, 150: 0.00015767026417652763,
+                                       151: 0.00010248567171474296, 152: 0.00029011328608481084,
+                                       153: 0.00023650539626479143, 154: 0.0005140050612154801,
+                                       155: 7.252832152120271e-05, 156: 6.779821359590688e-05,
+                                       157: 0.00013717312983357904, 158: 0.00014190323775887487,
+                                       159: 0.000268039449100097, 160: 0.0016539610712117748,
+                                       161: 0.00048247100838017453, 162: 0.0041025802738732485,
+                                       163: 0.0003452978785465955, 164: 0.0006590617042578855,
+                                       165: 0.0003799853366654316, 166: 0.0001324430219082832,
+                                       167: 0.0003106104204277594, 168: 0.00017816739851947623,
+                                       169: 0.00016240037210182346, 170: 7.410502416296798e-05,
+                                       171: 0.000302726907218933, 172: 0.00014505664304240542,
+                                       173: 0.0002333519909812609, 174: 0.0005392323034837245,
+                                       175: 0.0007725842944649854, 176: 0.0008608796424038409,
+                                       177: 0.00028853658344304555, 178: 0.00017974410116124149,
+                                       179: 0.0002916899887265761, 180: 0.00023650539626479143,
+                                       181: 0.004521983176582812, 182: 0.0040820831395303, 183: 0.00029011328608481084,
+                                       184: 0.0001734372905941804, 185: 0.0008656097503291366,
+                                       186: 0.00012140610341592628, 187: 0.00046828068460428706,
+                                       188: 0.00019866453286242482, 189: 0.0003011502045771678,
+                                       190: 0.00028380647551774974, 191: 0.000268039449100097,
+                                       192: 0.006237435650823433, 193: 0.0003058803125024636, 194: 0.004577167769044597,
+                                       195: 0.0001671304800271193, 196: 0.00018289750644477206,
+                                       197: 0.00013559642719181375, 198: 0.0007867746182408728,
+                                       199: 0.00018289750644477206, 200: 0.0002664627464583317,
+                                       201: 0.00035475809439718716, 202: 0.00025700253060774003,
+                                       203: 0.00039890576836661487, 204: 0.00022389177513066922,
+                                       205: 0.00012140610341592628, 206: 0.00010406237435650823,
+                                       207: 0.0013102398953069446, 208: 0.0003184939336365858,
+                                       209: 0.0008041183473002909, 210: 0.007708499215590436,
+                                       211: 0.0005124283585737148, 212: 0.00015767026417652763,
+                                       213: 0.0006685219201084772, 214: 0.00027749966495068865,
+                                       215: 0.0002664627464583317, 216: 0.00010879248228180406,
+                                       217: 0.00021443155928007758, 218: 0.0001340197245500485,
+                                       219: 9.775556378944712e-05, 220: 0.0007710075918232201,
+                                       221: 0.0002743462596671581, 222: 0.001414302269663453,
+                                       223: 0.00015609356153476234, 224: 0.007935544396004636,
+                                       225: 0.004510946258090455, 226: 0.00032795414948717745,
+                                       227: 0.00025542582796597474, 228: 0.0038676515802502226,
+                                       229: 0.0009901692590285934, 230: 0.0005849566800949175,
+                                       231: 0.000239658801548322, 232: 0.0008120018605091173,
+                                       233: 0.0001734372905941804, 234: 0.00025384912532420945,
+                                       235: 0.0003831387419489621, 236: 8.041183473002909e-05,
+                                       237: 0.00015451685889299708, 238: 0.0038881487145931712,
+                                       239: 0.00036264160760601354, 240: 0.00044147673969427735,
+                                       241: 9.617886114768185e-05, 242: 0.0039196827674284765,
+                                       243: 0.004255520430124481, 244: 0.0001324430219082832,
+                                       245: 0.00011509929284886517, 246: 8.987205058062074e-05,
+                                       247: 6.779821359590688e-05, 248: 0.0010185499065803684,
+                                       249: 0.0008671864529709019, 250: 0.00011825269813239572,
+                                       251: 0.00021127815399654703, 252: 0.00016870718266888456,
+                                       253: 0.0015388617783629096, 254: 0.0003405677706212997,
+                                       255: 0.003082453664651115}
+    LIKELYHOOD_EPILOG_BY_OPCODE_INT = {0: 0.0435043792915875, 1: 0.048012172144394426, 2: 0.012332968063887992,
+                                       3: 0.01824875637579131, 4: 0.005781768587353268, 5: 0.0016287338289435304,
+                                       6: 0.0007252832152120271, 7: 0.0011084219571609892, 8: 0.0014048420538128612,
+                                       9: 0.0004950846295142967, 10: 0.0052173090416012995, 11: 0.0004462068476195732,
+                                       12: 0.0004540903608283996, 13: 0.0005581527351849078, 14: 0.0005045448453648884,
+                                       15: 0.0005061215480066537, 16: 0.0008908369925973811, 17: 0.005352905468793113,
+                                       18: 0.0005660362483937342, 19: 0.00013717312983357904, 20: 0.005143204017438331,
+                                       21: 0.00847793010477189, 22: 0.01321434484063478, 23: 0.002152199106009602,
+                                       24: 0.0001923577222953637, 25: 0.00855834193950192, 26: 0.0005408090061254897,
+                                       27: 0.0005408090061254897, 28: 0.00032164733892011636, 29: 0.0002128548566383123,
+                                       30: 0.0038944555251602324, 31: 0.003736785260983705, 32: 0.03201337043840217,
+                                       33: 0.0004887778189472357, 34: 0.0041025802738732485, 35: 0.005146357422721862,
+                                       36: 0.0002743462596671581, 37: 0.003943333307054956, 38: 0.00019551112757889425,
+                                       39: 0.004011131520650863, 40: 0.0011431094152798253, 41: 0.00017816739851947623,
+                                       42: 0.0003705251208148399, 43: 0.00216954283506902, 44: 0.00015767026417652763,
+                                       45: 0.00027592296230892336, 46: 0.0003011502045771678,
+                                       47: 0.00042728641591838987, 48: 0.00422240967464741, 49: 0.0017643302561353442,
+                                       50: 0.005415973574463724, 51: 0.0010422004462068476, 52: 0.004181415405961513,
+                                       53: 0.0012692456266210474, 54: 0.008176779900194723, 55: 0.004996570671754161,
+                                       56: 0.0004146727947842677, 57: 0.004305974914660969, 58: 0.0009775556378944714,
+                                       59: 0.0005518459246178467, 60: 0.0004856244136637051, 61: 0.00016082366946005818,
+                                       62: 9.460215850591658e-05, 63: 0.00014505664304240542, 64: 0.00932619612604161,
+                                       65: 0.00018132080380300677, 66: 0.0012487484922780988,
+                                       67: 0.00014190323775887487, 68: 0.0008876835873138506,
+                                       69: 0.00022231507248890396, 70: 0.0005565760325431425, 71: 0.0005660362483937342,
+                                       72: 0.0004209796053513288, 73: 0.00044778355026133844,
+                                       74: 0.00024911901739891365, 75: 0.0005786498695278564,
+                                       76: 0.00024911901739891365, 77: 0.0011210355782951115,
+                                       78: 0.00022389177513066922, 79: 0.0009113341269403296, 80: 0.0787862543063691,
+                                       81: 0.010459845325470844, 82: 0.013053521171174722, 83: 0.0001340197245500485,
+                                       84: 0.007555559059339204, 85: 0.005138473909513035, 86: 0.028196173342688435,
+                                       87: 0.012974686039086459, 88: 0.00011509929284886517, 89: 0.004138844434633851,
+                                       90: 0.010858751093837458, 91: 0.03310917877442904, 92: 0.0002191616672053734,
+                                       93: 0.00010879248228180406, 94: 0.0002270451804141998,
+                                       95: 0.00032953085212894274, 96: 0.09900904238965053, 97: 0.025810622245697572,
+                                       98: 0.0016350406395105915, 99: 0.002614172980046828, 100: 0.0004209796053513288,
+                                       101: 0.0013023563820981183, 102: 0.00010406237435650823,
+                                       103: 0.0006984792703020174, 104: 0.001147839523205121,
+                                       105: 0.0010327402303562559, 106: 0.00024911901739891365,
+                                       107: 0.00037525522874013573, 108: 0.0002790763675924539,
+                                       109: 0.0003516046891136566, 110: 0.0005329254929166634,
+                                       111: 0.0005786498695278564, 112: 0.000376831931381901,
+                                       113: 0.0002711928543836275, 114: 0.000536078898200194, 115: 0.006016697280976294,
+                                       116: 0.00026961615174186227, 117: 0.0011967173050998446,
+                                       118: 0.000376831931381901, 119: 0.0003862921472324927,
+                                       120: 0.00020024123550419008, 121: 0.00019866453286242482,
+                                       122: 9.933226643121241e-05, 123: 0.0001718605879524151,
+                                       124: 0.0001135225902070999, 125: 0.0008025416446585256,
+                                       126: 0.0008671864529709019, 127: 0.0008608796424038409, 128: 0.0203741515368909,
+                                       129: 0.025413293179972724, 130: 0.015032282986590144, 131: 0.00976294275781059,
+                                       132: 0.001039047040923317, 133: 0.002762383028372764, 134: 0.0007268599178537924,
+                                       135: 0.006349381538388768, 136: 0.0042239863772891755, 137: 0.000119829400774161,
+                                       138: 0.000450936955544869, 139: 0.0006133373276466925,
+                                       140: 0.0004935079268725314, 141: 0.001078464606967449,
+                                       142: 0.00025069572004067893, 143: 0.0011336491994292336,
+                                       144: 0.047141832286139995, 145: 0.016681513949876624, 146: 0.006399836022925256,
+                                       147: 0.0032511608473199997, 148: 0.0008435359133444228,
+                                       149: 0.0016240037210182346, 150: 0.00029642009665187193,
+                                       151: 0.00025069572004067893, 152: 0.00016870718266888456,
+                                       153: 0.00036737171553130935, 154: 0.00010090896907297768,
+                                       155: 0.0009081807216567992, 156: 0.00014347994040064013,
+                                       157: 8.829534793885547e-05, 158: 0.0004540903608283996,
+                                       159: 0.00018132080380300677, 160: 0.0021884632667702035,
+                                       161: 0.002614172980046828, 162: 0.004126230813499728, 163: 0.0015782793444070417,
+                                       164: 0.0002065480460712512, 165: 0.0002459656121153831,
+                                       166: 0.00028222977287598445, 167: 0.0011099986598027546,
+                                       168: 0.0008529961291950145, 169: 0.0003705251208148399,
+                                       170: 0.0021080514320401746, 171: 0.00020339464078772065,
+                                       172: 0.00036106490496424826, 173: 0.00016397707474358872,
+                                       174: 0.0002995735019354025, 175: 0.0004209796053513288,
+                                       176: 0.003631146183985431, 177: 0.0004556670634701648,
+                                       178: 5.833799774531522e-05, 179: 0.0009822857458197671,
+                                       180: 0.00047301079252958287, 181: 0.00021600826192184284,
+                                       182: 0.0041025802738732485, 183: 0.0009507516929844616,
+                                       184: 0.0001340197245500485, 185: 0.0005486925193343161,
+                                       186: 0.0001466333456841707, 187: 0.0006133373276466925,
+                                       188: 0.00024754231475714836, 189: 0.00023808209890655672,
+                                       190: 0.00020970145135478175, 191: 0.00011667599549063044,
+                                       192: 0.0005628828431102036, 193: 0.0003389910679795344,
+                                       194: 0.0049035452158900095, 195: 0.00014347994040064013,
+                                       196: 0.001307086490023414, 197: 0.00012455950869945682,
+                                       198: 0.0010942316333851017, 199: 0.00011825269813239572,
+                                       200: 0.0004178262000677982, 201: 0.00016240037210182346,
+                                       202: 0.0005802265721696217, 203: 0.00012298280605769154,
+                                       204: 7.252832152120271e-05, 205: 0.0006953258650184869,
+                                       206: 0.0006890190544514257, 207: 0.0002049713434294859,
+                                       208: 0.00010721577964003879, 209: 0.0002932666913683414,
+                                       210: 0.0077384565657839756, 211: 0.00013559642719181375,
+                                       212: 0.00020339464078772065, 213: 0.00016397707474358872,
+                                       214: 0.00027749966495068865, 215: 0.0002979967992936372,
+                                       216: 0.00025384912532420945, 217: 0.0026567439513744905,
+                                       218: 0.0009539050982679922, 219: 0.00025069572004067893,
+                                       220: 0.00035475809439718716, 221: 0.0001340197245500485,
+                                       222: 0.0003121871230695247, 223: 6.937491623767216e-05,
+                                       224: 0.0009081807216567992, 225: 5.2031187178254115e-05,
+                                       226: 0.0008372291027773617, 227: 0.00025542582796597474,
+                                       228: 0.00390864584893612, 229: 0.000302726907218933, 230: 0.0003405677706212997,
+                                       231: 0.00015767026417652763, 232: 0.00040363587629191073,
+                                       233: 0.00025384912532420945, 234: 9.775556378944712e-05,
+                                       235: 0.00010406237435650823, 236: 0.00014821004832593596,
+                                       237: 0.00012928961662475266, 238: 0.003998517899516741,
+                                       239: 0.0009838624484615323, 240: 9.460215850591658e-05,
+                                       241: 0.0040127082232926286, 242: 0.005789652100562094,
+                                       243: 0.0059205184198286125, 244: 0.00019708783022065953,
+                                       245: 0.002235764346023162, 246: 0.00039417566044131906,
+                                       247: 0.00010563907699827352, 248: 0.0004714340898878176,
+                                       249: 0.0001718605879524151, 250: 0.0009744022326109408,
+                                       251: 0.0002522724226824442, 252: 0.0010295868250727255,
+                                       253: 0.0003043036098606983, 254: 0.0006101839223631619, 255: 0.07994040064014127}
+    AVERAGE_CONTRACT_SIZE = 1405
+    MIN_CONTRACT_SIZE = 6
+    MAX_CONTRACT_SIZE = 11526
+
+    def __init__(self, seed=None, length=None, prefix="0x"):
+        super().__init__(seed=seed)
+        self.length = length
+        self.prefix = prefix
+
+    def random_code_byte_sequence(self, length=None):
+        # todo: add gauss histogramm random.randgauss(min,max,avg) - triangle is not really correct here
+        length = length or int(random.triangular(self.MIN_CONTRACT_SIZE, 2 * self.AVERAGE_CONTRACT_SIZE + self.MIN_CONTRACT_SIZE))  # use gauss
+
+        rnd_prolog = WeightedRandomizer(self.LIKELYHOOD_PROLOG_BY_OPCODE_INT)
+        rnd_epilog = WeightedRandomizer(self.LIKELYHOOD_EPILOG_BY_OPCODE_INT)  # not completely true as this incorps. pro/epilog
+        rnd_corpus = WeightedRandomizer(self.LIKELYHOOD_BY_OPCODE_INT)
+
+        b = []
+        for _ in range(128):
+            b.append(rnd_prolog.random())
+        for _ in range(length - 128 * 2):
+            b.append(rnd_corpus.random())
+        for _ in range(128):
+            b.append(rnd_epilog.random())
+
+        return bytes(b)
+
+    def generate(self, length=50):
+        return "%s%s" % (self.prefix,
+                         binascii.hexlify(self.random_code_byte_sequence(self.length)).decode("utf-8"))
+
